mudança do tipo do atributo valor do Vertex para Object e do atributo distance do Edge para int

+++ Java/ArvoreGenerica/src/main/java/ifrn/tads/estruturadedados/graph/EdgeInterface.java
@@ -39,10 +39,10 @@ public interface EdgeInterface {
     /**
      * @return the label
      */
-    double getDistance();
+    int getDistance();
 
     /**
      * @param distance the distance to set
      */
-    void setDistance(double distance);
+    void setDistance(int distance);
 }
